stateflow implementation and observation example
