+++ app/src/main/java/com/vu/androidbasicapp/ui/home/HomeScreenFragment.kt
@@ -1,0 +1,41 @@
+package com.vu.androidbasicapp.ui.home
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
+import androidx.lifecycle.lifecycleScope
+import com.vu.androidbasicapp.R
+import kotlinx.coroutines.launch
+
+
+class HomeScreenFragment : Fragment() {
+
+    private val viewModel: HomeScreenViewModel by viewModels()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_home_screen, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        lifecycleScope.launch {
+            viewModel.greetingText.collect { greetingTextState ->
+                view.findViewById<TextView>(R.id.screenTitle).text = greetingTextState
+            }
+        }
+    }
+}
