+++ app/src/main/java/com/vu/androidbasicapp/ui/home/HomeScreenViewModel.kt
@@ -1,0 +1,23 @@
+package com.vu.androidbasicapp.ui.home
+
+import android.util.Log
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.launch
+
+class HomeScreenViewModel: ViewModel() {
+
+    val greetingText = MutableStateFlow("Hello Class")
+
+    init {
+        Log.d("nit3213", "HomeScreenViewModel ViewModel injected ")
+
+        viewModelScope.launch {
+            delay(4000)
+            greetingText.value = "Hi the greeting has changed!"
+        }
+    }
+}
