+++ app/src/main/res/layout/fragment_home_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.HomeScreenFragment">
+    tools:context=".ui.home.HomeScreenFragment">
 
     <TextView
         android:id="@+id/screenTitle"
